Adding to AvgDailyCalories Method
I found a simpler way, I believe to build a unique array of dates to
work with in this function. I am quite please with the technique and
will use this as my go to method going forward. It start with a blank
array, and then looks to see if the first spot of the array we are
checking is present in our new blank array, it will, of course, not be
present(i make the check with indexOf), so it is pushed. As the loop
executes it only pushes the value of the array being read if the return
from indexOf is -1 (not present). I think this is some what easier for
this task but won't really give a word count.